requirements.txt
pulsar-client==2.6.1